// backend-dex-api-server-feature-two-way-orders/src/implementation/client/sumsub/api.ts
export interface SumsubAccessToken {
    /**
     * 
     * @type {string}
     * @memberof SumsubAccessToken
     */
    'token': string;
}
/**
 * 
 * @export
 * @interface SumsubApplicant
 */
export interface SumsubApplicant {
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicant
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicant
     */
    'createDate': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicant
     */
    'reviewDate': string;
    /**
     * 
     * @type {SumsubApplicantInfo}
     * @memberof SumsubApplicant
     */
    'info': SumsubApplicantInfo;
    /**
     * 
     * @type {SumsubReviewResult}
     * @memberof SumsubApplicant
     */
    'reviewStatus'?: SumsubReviewResult;
}
// ...
export interface SumsubApplicantInfo {
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantInfo
     */
    'country': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantInfo
     */
    'nationality': string;
    /**
     * 
     * @type {Array<SumsubIdentityDocument>}
     * @memberof SumsubApplicantInfo
     */
    'idDocs'?: Array<SumsubIdentityDocument>;
}
/**
 * 
 * @export
 * @interface SumsubApplicantReference
 */
export interface SumsubApplicantReference {
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantReference
     */
    'applicantId': string;
}
/**
 * 
 * @export
 * @interface SumsubApplicantStatus
 */
export interface SumsubApplicantStatus {
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantStatus
     */
    'applicantId': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantStatus
     */
    'levelName'?: string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantStatus
     */
    'createDate': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantStatus
     */
    'reviewDate': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubApplicantStatus
     */
    'reviewStatus': string;
    /**
     * 
     * @type {SumsubReviewResult}
     * @memberof SumsubApplicantStatus
     */
    'reviewResult'?: SumsubReviewResult;
}
// ...
export interface SumsubCreateApplicant {
    /**
     * 
     * @type {string}
     * @memberof SumsubCreateApplicant
     */
    'sourceKey': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubCreateApplicant
     */
    'externalUserId': string;
}
/**
 * 
 * @export
 * @interface SumsubIdentityDocument
 */
export interface SumsubIdentityDocument {
    /**
     * 
     * @type {string}
     * @memberof SumsubIdentityDocument
     */
    'idDocType': string;
    /**
     * 
     * @type {string}
     * @memberof SumsubIdentityDocument
     */
    'validUntil': string;
}
/**
 * 
 * @export
 * @interface SumsubReviewResult
 */
export interface SumsubReviewResult {
    /**
     * 
     * @type {string}
     * @memberof SumsubReviewResult
     */
    'reviewAnswer': string;
}
// ...
export class DefaultApi extends BaseAPI {
    /**
     * 
     * @param {string} userId ID of an user
     * @param {string} levelName access token level name
     * @param {number} ttlInSecs access token time to live in seconds
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public createAccessToken(userId: string, levelName: string, ttlInSecs: number, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).createAccessToken(userId, levelName, ttlInSecs, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} levelName access token level name
     * @param {SumsubCreateApplicant} [sumsubCreateApplicant] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public createApplicant(levelName: string, sumsubCreateApplicant?: SumsubCreateApplicant, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).createApplicant(levelName, sumsubCreateApplicant, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Array<string>} applicantId ID of an applicant
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getApplicant(applicantId: Array<string>, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getApplicant(applicantId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} userId ID of an user
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getApplicantByUserId(userId: string, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getApplicantByUserId(userId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Array<string>} applicantId ID of an applicant
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getApplicantStatus(applicantId: Array<string>, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getApplicantStatus(applicantId, options).then((request) => request(this.axios, this.basePath));
    }
}